requirements.txt
openai
pinecone[grpc]
sentence-transformers
transformers
accelerate
datasets
langchain
langchain_community
langchain-huggingface
langchain-openai
tiktoken
python-dotenv
rank_bm25
einops
altair<5
aiofiles==23.2.1
annotated-types==0.6.0
anyio==4.3.0
fastapi==0.110.1
pydantic==2.7.0
pydantic_core==2.18.1
python-jose==3.3.0
python-multipart==0.0.9
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
uvicorn==0.29.0